+++ idea/testData/refactoring/rename/renameKotlinFacadeClassWithDefaultNameAndCommentsByRef/after/test/JavaClient.java
@@ -1,0 +1,7 @@
+package test;
+
+class Test {
+    {
+        Foo.foo();
+    }
+}