+++ idea/testData/refactoring/rename/renameKotlinFacadeClassWithDefaultNameByRef/after/test/test.kt
@@ -1,0 +1,7 @@
+@file:JvmName("Foo")
+
+package test
+
+fun foo() {
+
+}