+++ idea/testData/refactoring/rename/renameKotlinFunWithJvmName/after/test/test.kt
@@ -1,0 +1,10 @@
+package test
+
+class A {
+    @JvmName("foo")
+    fun second() = 1
+}
+
+fun test() {
+    A().second()
+}