+++ idea/testData/refactoring/rename/renameKotlinFunWithJvmNameByJavaRef/before/test/JavaClient.java
@@ -1,0 +1,7 @@
+package test;
+
+class Test {
+    {
+        new A()./*rename*/foo();
+    }
+}