+++ idea/testData/refactoring/rename/renameKotlinFunWithJvmNameToDefaultByJavaRef/before/test/test.kt
@@ -1,0 +1,10 @@
+package test
+
+class A {
+    @JvmName("foo")
+    fun first() = 1
+}
+
+fun test() {
+    A().first()
+}