+++ idea/testData/refactoring/rename/renameKotlinPackageClass/before/testing/JavaClient.java
@@ -1,9 +1,0 @@
-package testing;
-
-import testing.rename.RenameKotlinClassKt;
-
-class JavaClient {
-    void foo() {
-        RenameKotlinClassKt.test();
-    }
-}