+++ idea/testData/refactoring/rename/renameKotlinParameterWithSetterJvmName/after/test/JavaClient.java
@@ -1,0 +1,8 @@
+package test;
+
+class Test {
+    {
+        new A().getSecond();
+        new A().setBar(1);
+    }
+}