+++ idea/testData/refactoring/rename/renameKotlinPropertyWithGetterJvmName/after/test/test.kt
@@ -1,0 +1,11 @@
+package test
+
+class A {
+    @get:JvmName("getFoo")
+    var second = 1
+}
+
+fun test() {
+    A().second
+    A().second = 1
+}