+++ idea/testData/refactoring/rename/renameKotlinPropertyWithSetterJvmNameByGetterRef/after/test/test.kt
@@ -1,0 +1,11 @@
+package test
+
+class A {
+    @set:JvmName("setBar")
+    var second = 1
+}
+
+fun test() {
+    A().second
+    A().second = 1
+}